--- common/src/main/java/com/alibaba/datax/common/plugin/AbstractTaskPlugin.java
@@ -1,37 +1,0 @@
-package com.alibaba.datax.common.plugin;
-
-/**
- * Created by jingxing on 14-8-24.
- */
-public abstract class AbstractTaskPlugin extends AbstractPlugin {
-
-    //TaskPlugin 应该具备taskId
-    private int taskGroupId;
-    private int taskId;
-    private TaskPluginCollector taskPluginCollector;
-
-    public TaskPluginCollector getTaskPluginCollector() {
-        return taskPluginCollector;
-    }
-
-    public void setTaskPluginCollector(
-            TaskPluginCollector taskPluginCollector) {
-        this.taskPluginCollector = taskPluginCollector;
-    }
-
-    public int getTaskId() {
-        return taskId;
-    }
-
-    public void setTaskId(int taskId) {
-        this.taskId = taskId;
-    }
-
-    public int getTaskGroupId() {
-        return taskGroupId;
-    }
-
-    public void setTaskGroupId(int taskGroupId) {
-        this.taskGroupId = taskGroupId;
-    }
-}
