--- common/src/main/java/com/alibaba/datax/common/plugin/PluginCollector.java
@@ -1,9 +1,0 @@
-package com.alibaba.datax.common.plugin;
-
-
-/**
- * 这里只是一个标示类
- * */
-public interface PluginCollector {
-
-}
