--- datax-web/src/main/assembly/package.xml
@@ -1,44 +1,0 @@
-<assembly
-	xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0 http://maven.apache.org/xsd/assembly-1.1.0.xsd">
-	<id>web</id>
-	<formats>
-		<format>dir</format>
-	</formats>
-	<includeBaseDirectory>false</includeBaseDirectory>
-	<fileSets>
-		<!--<fileSet>-->
-			<!--<directory>src/main/resources</directory>-->
-			<!--<includes>-->
-				<!--<include>plugin.json</include>-->
-				<!--<include>plugin_job_template.json</include>-->
-			<!--</includes>-->
-			<!--<outputDirectory>plugin/reader/ftpreader</outputDirectory>-->
-		<!--</fileSet>-->
-		<fileSet>
-			<directory>target/</directory>
-			<includes>
-				<include>datax-web-0.0.1-SNAPSHOT.jar</include>
-			</includes>
-			<outputDirectory>plugin/web</outputDirectory>
-		</fileSet>
-		<!--<fileSet>-->
-		<!--<directory>src/main/lib</directory>-->
-		<!--<includes>-->
-		<!--<include>ojdbc6-11.2.0.3.jar</include>-->
-		<!--<include>sqljdbc4-4.0.jar</include>-->
-		<!--</includes>-->
-		<!--<outputDirectory>plugin/writer/oraclewriter/libs</outputDirectory>-->
-		<!--</fileSet>-->
-	</fileSets>
-
-	<!--该模块的依赖和实体datax-web-0.0.1-SNAPSHOT.jar分开存放-->
-	<!--<dependencySets>-->
-	<!--<dependencySet>-->
-	<!--<useProjectArtifact>false</useProjectArtifact>-->
-	<!--<outputDirectory>plugin/web/libs</outputDirectory>-->
-	<!--<scope>runtime</scope>-->
-	<!--</dependencySet>-->
-	<!--</dependencySets>-->
-</assembly>
