+++ datax-web/src/main/java/com/wugui/dataxweb/service/impl/IDataxJobServiceImpl.java
@@ -2,18 +2,18 @@ package com.wugui.dataxweb.service.impl;
 
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.util.StrUtil;
-import com.alibaba.datax.common.log.EtlJobFileAppender;
-import com.alibaba.datax.common.log.EtlJobLogger;
-import com.alibaba.datax.common.log.LogResult;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import com.wugui.dataxweb.dto.RunJobDto;
 import com.wugui.dataxweb.entity.JobLog;
+import com.wugui.dataxweb.log.CollectingLogOutputStream;
+import com.wugui.dataxweb.log.EtlJobFileAppender;
+import com.wugui.dataxweb.log.EtlJobLogger;
+import com.wugui.dataxweb.log.LogResult;
 import com.wugui.dataxweb.service.IDataxJobService;
 import com.wugui.dataxweb.service.IJobLogService;
-import com.wugui.dataxweb.util.CollectingLogOutputStream;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.exec.CommandLine;
 import org.apache.commons.exec.DefaultExecuteResultHandler;
