+++ datax-web/src/main/java/com/wugui/tool/util/DasUtil.java
@@ -1,9 +1,9 @@
 package com.wugui.tool.util;
 
 import com.wugui.dataxweb.entity.JobJdbcDatasource;
-import com.wugui.tool.database.TableInfo;
 import com.wugui.tool.query.BaseQueryTool;
 import com.wugui.tool.query.QueryToolFactory;
+import database.TableInfo;
 import lombok.extern.slf4j.Slf4j;
 
 /**
