+++ pom.xml
@@ -1,52 +1,58 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.alibaba.datax</groupId>
+    <groupId>com.wugui</groupId>
     <artifactId>datax-all</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <dependencies>
-        <dependency>
-            <groupId>org.hamcrest</groupId>
-            <artifactId>hamcrest-core</artifactId>
-            <version>1.3</version>
-        </dependency>
-    </dependencies>
-
+    <version>1.0.0</version>
     <name>datax-all</name>
     <packaging>pom</packaging>
 
+
     <properties>
-        <datax-project-version>0.0.1-SNAPSHOT</datax-project-version>
-        <commons-lang3-version>3.3.2</commons-lang3-version>
-        <commons-configuration-version>1.10</commons-configuration-version>
-        <commons-cli-version>1.4</commons-cli-version>
-        <fastjson-version>1.1.46.sec01</fastjson-version>
-        <guava-version>16.0.1</guava-version>
-        <diamond.version>3.7.2.1-SNAPSHOT</diamond.version>
-
-        <!--slf4j 1.7.10 和 logback-classic 1.0.13 是好基友 -->
-        <slf4j-api-version>1.7.10</slf4j-api-version>
-        <logback-classic-version>1.0.13</logback-classic-version>
-        <commons-io-version>2.4</commons-io-version>
-        <junit-version>4.11</junit-version>
-        <tddl.version>5.1.22-1</tddl.version>
-        <swift-version>1.0.0</swift-version>
-
-
         <project-sourceEncoding>UTF-8</project-sourceEncoding>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <maven.compiler.encoding>UTF-8</maven.compiler.encoding>
+
+        <datax-project-version>0.0.1-SNAPSHOT</datax-project-version>
+        <commons-lang3-version>3.3.2</commons-lang3-version>
+        <commons-configuration-version>1.10</commons-configuration-version>
+        <commons-cli-version>1.4</commons-cli-version>
+        <fastjson-version>1.2.62</fastjson-version>
+        <guava-version>28.1-jre</guava-version>
+        <diamond.version>3.7.2.1-SNAPSHOT</diamond.version>
+
+        <!--slf4j 1.7.10 和 logback-classic 1.0.13 是好基友 -->
+        <slf4j-api-version>1.7.10</slf4j-api-version>
+        <logback-classic-version>1.2.2</logback-classic-version>
+        <commons-io-version>2.4</commons-io-version>
+        <junit-version>4.11</junit-version>
+        <tddl.version>5.1.22-1</tddl.version>
+        <swift-version>1.0.0</swift-version>
+        <commons-exec-version>1.3</commons-exec-version>
+        <hutool.version>4.5.1</hutool.version>
+        <datax.version>1.0.0</datax.version>
+        <hamcrest.version>1.3</hamcrest.version>
+        <mockito.version>1.9.5</mockito.version>
+        <postgresql.version>42.2.5</postgresql.version>
+        <druid.version>1.1.14</druid.version>
+        <mysql-connector.version>5.1.47</mysql-connector.version>
     </properties>
 
     <modules>
-        <module>common</module>
-
         <!--springboot-web-->
         <module>datax-web</module>
-
+        <module>datax-common</module>
     </modules>
+
+    <parent>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-starter-parent</artifactId>
+        <version>2.1.4.RELEASE</version>
+        <relativePath/> <!-- lookup parent from repository -->
+    </parent>
 
     <dependencyManagement>
         <dependencies>
@@ -60,11 +66,11 @@
                 <artifactId>fastjson</artifactId>
                 <version>${fastjson-version}</version>
             </dependency>
-            <!--<dependency>
+            <dependency>
                 <groupId>com.google.guava</groupId>
                 <artifactId>guava</artifactId>
                 <version>${guava-version}</version>
-            </dependency>-->
+            </dependency>
             <dependency>
                 <groupId>commons-io</groupId>
                 <artifactId>commons-io</artifactId>
@@ -118,39 +124,57 @@
             <dependency>
                 <groupId>org.mockito</groupId>
                 <artifactId>mockito-all</artifactId>
-                <version>1.9.5</version>
-                <scope>test</scope>
+                <version>${mockito.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.hamcrest</groupId>
+                <artifactId>hamcrest-core</artifactId>
+                <version>${hamcrest.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.wugui</groupId>
+                <artifactId>datax-web</artifactId>
+                <version>${datax.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.wugui</groupId>
+                <artifactId>datax-common</artifactId>
+                <version>${datax.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.apache.commons</groupId>
+                <artifactId>commons-exec</artifactId>
+                <version>${commons-exec-version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>cn.hutool</groupId>
+                <artifactId>hutool-all</artifactId>
+                <version>${hutool.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.postgresql</groupId>
+                <artifactId>postgresql</artifactId>
+                <version>${postgresql.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.alibaba</groupId>
+                <artifactId>druid</artifactId>
+                <version>${druid.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>mysql</groupId>
+                <artifactId>mysql-connector-java</artifactId>
+                <version>${mysql-connector.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
 
-    <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-assembly-plugin</artifactId>
-                <configuration>
-                    <finalName>datax</finalName>
-                    <descriptors>
-                        <descriptor>package.xml</descriptor>
-                    </descriptors>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>make-assembly</id>
-                        <phase>package</phase>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>2.3.2</version>
-                <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
-                    <encoding>${project-sourceEncoding}</encoding>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
 </project>
